Extract test data to reuse
